The modules can now also be registered and unregistered. Still, the table is NOT YET used (still using config.ini)
The order is relevant, as the settings are processed in sequence

This should work fine when the extension has no settings.

When an extension is UNregistered, all related settings are also deleted. The extension_id is an auto-increment, and should never re-occur.

If there are any settings for an extension, then (TODO):

*) enabling the extension also requires the mandatory settings (if any)
*) Extra action to change settings, once the extension is enabled

Marcel