Typo in translation.
In init, $config is made rw (option 'true' added)
The last 5 lines will scan the DB, and create the extensions part of $config  

WARNING: As the extensions and variable modifications are implemented as a plugin, you will loose this functionality when you disabled the 'default invoices' plugin! (actually, the web interface to it)

You can enable/disable the tab-menu, to see things working.
There is an issue with dbQuery though (see forum). That is why domain_id = 1 is hardcoded in the SQL